Switch from Float to Double to pass test

+++ src/GoldenSection.hs
@@ -15,11 +15,11 @@ invphi2 = (3 - sqrt 5) / 2
 -- | Golden section search, recursive.
 -- Given a function f with a single local maximum in the interval [a, b], golden section search returns a subset interval [c, d] that contains the maximum with d-c <= tolerance
 goldenSectionSearch ::
-    (Float -> Float) -- ^ Function with a single maximum in [a, b]
-    -> Float -- ^ One side of the interval
-    -> Float -- ^ Other side of the interval
-    -> Float -- ^ Tolerance
-    -> (Float, Float) -- ^ The interval in which the maximum is
+    (Double -> Double) -- ^ Function with a single maximum in [a, b]
+    -> Double -- ^ One side of the interval
+    -> Double -- ^ Other side of the interval
+    -> Double -- ^ Tolerance
+    -> (Double, Double) -- ^ The interval in which the maximum is
 
 -- Just a wrapper function because of all the ugly Nothing's of the recursive function
 goldenSectionSearch f a b tolerance = goldenSectionSearchRecursive f a b tolerance Nothing Nothing Nothing Nothing Nothing
@@ -28,17 +28,18 @@ goldenSectionSearch f a b tolerance = goldenSectionSearchRecursive f a b toleran
 -- Given a function f with a single local maximum in the interval [a, b], golden section search returns a subset interval [c, d] that contains the maximum with d-c <= tolerance
 -- Taken from the python implementation at https://en.wikipedia.org/wiki/Golden-section_search
 goldenSectionSearchRecursive ::
-    (Float -> Float) -- ^ Function with a single maximum in [a, b]
-    -> Float -- ^ One side of the interval
-    -> Float -- ^ Other side of the interval
-    -> Float -- ^ Tolerance
-    -> Maybe Float -- ^ h, Current search interval
-    -> Maybe Float -- ^ c, New left interval point. If Nothing, a new point is chosen.
-    -> Maybe Float -- ^ d, New right interval point.
-    -> Maybe Float -- ^ f(c), Function value at c
-    -> Maybe Float -- ^ f(d), Function value at d
-    -> (Float, Float) -- ^ The interval in which the maximum is
+    (Double -> Double) -- ^ Function with a single maximum in [a, b]
+    -> Double -- ^ One side of the interval
+    -> Double -- ^ Other side of the interval
+    -> Double -- ^ Tolerance
+    -> Maybe Double -- ^ h, Current search interval
+    -> Maybe Double -- ^ c, New left interval point. If Nothing, a new point is chosen.
+    -> Maybe Double -- ^ d, New right interval point.
+    -> Maybe Double -- ^ f(c), Function value at c
+    -> Maybe Double -- ^ f(d), Function value at d
+    -> (Double, Double) -- ^ The interval in which the maximum is
 
+-- For debugging, prints all function calls
 --goldenSectionSearchRecursive f a' b' tolerance h' c' d' fc' fd' | trace ("myfun a=" ++ show a' ++ " b=" ++ show b' ++ " h=" ++ show h' ++ " c=" ++ show c' ++ " d=" ++ show d' ++ " fc=" ++ show fc' ++ " fd=" ++ show fd') False = undefined
 goldenSectionSearchRecursive f a' b' tolerance h' c' d' fc' fd'
     | h < tolerance = (a, b)
